performance improve
git-svn-id: file:///vagrant/svn/dsis@13 6a918f4f-bbb9-b144-964e-14738c76912e

+++ trunk/src/Core/src/Util/CountEnumerable.cs
@@ -13,10 +13,10 @@ namespace DSIS.Core.Util
     private readonly IEnumerable<T> myEnumerable;
     public readonly int Count;
 
-    public CountEnumerable(IEnumerable<T> enumerable, int count)
+    public CountEnumerable(IEnumerable<T> enumerable, long count)
     {
       myEnumerable = enumerable;
-      Count = count;
+      Count = (int)count;
     }
 
     IEnumerator<T> IEnumerable<T>.GetEnumerator()
